Move Signature to Common
