+++ kifi-backend/modules/common/app/com/keepit/rover/article/Signature.scala
@@ -1,7 +1,11 @@
-package com.keepit.scraper
+package com.keepit.rover.article
 
 import java.util.Arrays
 import javax.xml.bind.DatatypeConverter._
+import org.apache.lucene.analysis.standard.StandardTokenizer
+import org.apache.lucene.analysis.tokenattributes.CharTermAttribute
+import java.io.StringReader
+
 import scala.math._
 
 class Signature(val bytes: Array[Byte]) {
@@ -21,8 +25,8 @@ class Signature(val bytes: Array[Byte]) {
 
 object Signature {
 
-  import play.api.libs.json._
   import play.api.libs.functional.syntax._
+  import play.api.libs.json._
 
   implicit val format: Format[Signature] = Format[Signature](Reads.of[String].fmap(Signature.apply), Writes { signature => JsString(signature.toBase64()) })
 
@@ -104,4 +108,21 @@ object Signature {
       }
     }
   }
+}
+
+class SignatureBuilder(windowSize: Int = 20) extends Signature.Builder(windowSize) {
+
+  override protected def tokenize(text: String)(addTerm: (Array[Char], Int) => Unit): Unit = {
+    val ts = new StandardTokenizer(new StringReader(text))
+    val termAttr = ts.getAttribute(classOf[CharTermAttribute])
+
+    try {
+      ts.reset()
+      while (ts.incrementToken()) addTerm(termAttr.buffer(), termAttr.length())
+      ts.end()
+    } finally {
+      ts.close()
+    }
+  }
+
 }
