+++ kifi-backend/modules/common/app/com/keepit/shoebox/ShoeboxScraperClient.scala
@@ -9,8 +9,9 @@ import com.keepit.common.routes.Shoebox
 import com.keepit.common.service.ThrottledServiceClient
 import com.keepit.common.zookeeper.ServiceCluster
 import com.keepit.model._
+import com.keepit.rover.article.Signature
 import com.keepit.rover.fetcher.HttpRedirect
-import com.keepit.scraper.{ Signature, ScrapeRequest }
+import com.keepit.scraper.ScrapeRequest
 import org.joda.time.DateTime
 import play.api.libs.json.{ JsString, JsArray, Json }
 import play.api.libs.json.Json.JsValueWrapper
