+++ kifi-backend/modules/shoebox/app/com/keepit/normalizer/NormalizationCandidate.scala
@@ -1,9 +1,9 @@
 package com.keepit.normalizer
 
 import com.keepit.model.{ NormalizedURI, RawKeep, Normalization }
-import com.keepit.scraper.Signature
 import com.keepit.common.db.Id
 import com.keepit.commanders.RawBookmarkRepresentation
+import com.keepit.rover.article.Signature
 import play.api.libs.json._
 import play.api.libs.functional.syntax._
