+++ kifi-backend/project/Global.scala
@@ -81,7 +81,8 @@ object Global {
     "com.mchange" % "c3p0" % "0.9.5-pre8", // todo(andrew): remove from common when C3P0 plugin is in sqldb
     "org.im4java" % "im4java" % "1.4.0", //todo(eishay): means that all services get that. not sure that's this is best
     "org.apache.httpcomponents" % "httpclient" % "4.3.1", // todo(Léo): remove when Scraper is gone and ApacheHttpFetcher is in Rover
-    "org.apache.tika" % "tika-parsers" % "1.5" // todo(Léo): remove when Scraper is gone and Tika is in Rover
+    "org.apache.tika" % "tika-parsers" % "1.5", // todo(Léo): remove when Scraper is gone and Tika is in Rover,
+    "org.apache.lucene" % "lucene-analyzers-common" % "4.10.2" // todo(Léo): remove when Scraper is gone and Signature is in Rover,
   ) map (_.excludeAll(
     ExclusionRule(organization = "javax.jms"),
     ExclusionRule(organization = "com.sun.jdmk"),
